refactor: update supabase client imports
